Add enum_help gem and translations
