--- app/views/expenses/_detail.html.erb
@@ -1,4 +1,0 @@
-<li><%= t("expenses.defaults.name") %>: <%= expense.name %></li>
-<li><%= t("expenses.defaults.due_date") %>: <%= expense.due_date.strftime("%d-%m-%Y") %></li>
-<li><%= t("expenses.defaults.frequency") %>: <%= expense.frequency %></li>
-<li><%= t("expenses.defaults.bill_estimate") %>: <%= number_to_currency(expense.bill_estimate) %></li>
