+++ app/views/expenses/edit.html.erb
@@ -1,6 +1,2 @@
 <h1><%= t(".title") %></h1>
-<%= simple_form_for expense do |f| %>
-  <%= render "form", f: f %>
-<% end %>
-
-<%= link_to t(".delete"), expense, method: :delete %>
+  <%= render "form" %>
