+++ app/views/expenses/index.html.erb
@@ -1,8 +1,11 @@
-<div><%= t(".title") %></div>
-
+<h1><%= t(".title") %></h1>
+<% @expenses.each do |expense| %>
 <ul>
-  <% @expenses.each do |expense| %>
-    <%= render "detail", expense: expense %>
-    </br>
-  <% end %>
+  <li><%= t("expenses.defaults.name") %>: <%= expense.name %></li>
+  <li><%= t("expenses.defaults.due_date") %>: <%= l expense.due_date, format: :default %></li>
+  <li><%= t("expenses.defaults.frequency") %>: <%= expense.frequency %></li>
+  <li><%= t("expenses.defaults.bill_estimate") %>: <%= number_to_currency(expense.bill_estimate) %></li>
+  <%= link_to t(".delete"), expense_path(expense), method: :delete %> |
+  <%= link_to t(".update"), edit_expense_path(expense) %>
 </ul>
+<% end %>
