+++ app/views/expenses/new.html.erb
@@ -1,4 +1,2 @@
 <h1><%= t(".title") %></h1>
-<%= simple_form_for @expense do |f| %>
-  <%= render "form", f: f %>
-<% end %>
+  <%= render "form" %>
