+++ app/views/expenses/show.html.erb
@@ -1,5 +1,8 @@
-<div><%= t(".title", name: expense.name) %></div>
-
+<h1><%= t(".title") %></h1>
 <ul>
-  <%= render "detail" %>
+  <li><%= t("expenses.defaults.name") %>: <%= expense.name %></li>
+  <li><%= t("expenses.defaults.due_date") %>: <%= l expense.due_date, format: :default %></li>
+  <li><%= t("expenses.defaults.frequency") %>: <%= expense.frequency %></li>
+  <li><%= t("expenses.defaults.bill_estimate") %>: <%= number_to_currency(expense.bill_estimate) %></li>
+  <li><%= t("expenses.defaults.budget") %>: <%= @budget.name %></li>
 </ul>
